optimisation: set enum types explictly to uint8_t
This commit produces the same savings as the compiler options -fshort-enums. Except by setting the types manually we save also 2 bytes of SRAM.

By default, the enum type is 2 bytes, with we can explictly set it to one byte when applicable to reduce code size.

Almost all the savings from from 'enum Mode' in leds.h.

Change in memory:
Flash: -116 bytes
SRAM: -2 bytes

+++ src/logic/result_codes.h
@@ -7,7 +7,7 @@
 /// therefore the error codes have been extracted to one place.
 ///
 /// Please note that currently only LoadFilament can return something else than "OK"
-enum class ResultCode : uint_fast16_t {
+enum class ResultCode : uint_fast8_t {
     OK = 0,
     Cancelled = 1
 };
